backend/requirements.txt
python-decouple
python-multipart
deta
fastapi
fastapi_mail
fastapi_utils
passlib
bcrypt
PyJWT
uvicorn

